Backend/requirements.txt
asgiref==3.4.1
boto3==1.19.6
botocore==1.22.6
Django==3.2.8
django-storages==1.12.2
djangorestframework==3.12.4
jmespath==0.10.0
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
